Added food, chores and general viewcontroller
I added a food, chores and general viewcontroller while also creating the ability for navigation though the viewcontrollers. Now you can tap on a button on the main screen, and the viewcontroller will switch. You can also press the 'back' button to go back to the main menu. There is also a new 'Go to Schedule' button which does not do anything yet but will be connected to the schedule viewcontroller.